Refactor the entire codebase to match the new syntax for Gravity
