+++ src/actions/resize/ResizeActions/shared/ResizeAdvancedAction.ts
@@ -1,5 +1,5 @@
 import ResizeSimpleAction from "./ResizeSimpleAction";
-import {GravityQualifier} from "../../../../values/gravity/Gravity";
+import {IAnyGravity} from "../../../../values/gravity/GravityQualifier";
 
 /**
  * @description Defines an advanced resize.
@@ -11,7 +11,7 @@ class ResizeAdvancedAction extends ResizeSimpleAction {
    * @description Which part of the original image to include.
    * @param {Values.Gravity} gravity
    */
-  gravity(gravity:GravityQualifier): this {
+  gravity(gravity:IAnyGravity): this {
     return this.addQualifier(gravity);
   }
 }
